+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.3
-*
+* Change test lib to buster-node + referee
 
 ### 0.0.2
 * Change initLocales signature to accept base directory